end project

+++ controller/product.controller.js
@@ -44,17 +44,27 @@ exports.findByCatagory = async (req, res) => {
 
 exports.uploadProduct = async (req, res) => {
     try {
-        const newProduct = new Product({
-            id: uuidv4(),
-            title: req.body.title,
-            catagory: req.body.catagory,
-            price: Number(req.body.price),
-            rating: Number(req.body.rating),
-            description: req.body.description,
-            photoUrl: req.body.photoUrl,
-        })
-        await newProduct.save();
-        res.status(200).json({ message: 'successfully created Product', newProduct });
+        const requesterEmail = req.decodedEmail;
+        if (requesterEmail) {
+            const requesterAccount = await User.findOne({
+                email: requesterEmail,
+            });
+            if (requesterAccount.role === 'admin' || requesterAccount.role === 'administer') {
+                const newProduct = new Product({
+                    id: uuidv4(),
+                    title: req.body.title,
+                    catagory: req.body.catagory,
+                    price: Number(req.body.price),
+                    rating: Number(req.body.rating),
+                    description: req.body.description,
+                    photoUrl: req.body.photoUrl,
+                })
+                await newProduct.save();
+                res.status(200).json({ message: 'successfully created Product', newProduct });
+            } else {
+                res.status(500).json({ message: 'Only Admin can upload product' });
+            }
+        }
     } catch (error) {
         res.status(500).json(error.message);
     }
@@ -62,28 +72,46 @@ exports.uploadProduct = async (req, res) => {
 
 exports.updateProduct = async (req, res) => {
     try {
-        const product = await Product.findOne({ id: req.params.id });
-        product.title = req.body.title;
-        product.price = req.body.price;
-        product.description = req.body.description;
-        product.rating = req.body.rating;
-        product.catagory = req.body.catagory;
-        await product.save();
-        res.status(200).json(product);
+        const requesterEmail = req.decodedEmail;
+        if (requesterEmail) {
+            const requesterAccount = await User.findOne({
+                email: requesterEmail,
+            });
+            if (requesterAccount.role === 'admin' || requesterAccount.role === 'administer') {
+                const product = await Product.findOne({ id: req.params.id });
+                product.title = req.body.title;
+                product.price = req.body.price;
+                product.description = req.body.description;
+                product.catagory = req.body.catagory;
+                product.photoUrl = req.body.photoUrl;
+                await product.save();
+                res.status(200).json(product);
+            } else {
+                res.status(500).json({ message: 'Only Admin can update product' });
+
+            }
+        }
     } catch (error) {
         res.status(500).json(error.message);
     }
 }
 
 exports.deleteProduct = async (req, res) => {
-
     try {
-        await Product.deleteOne({ id: req.params.id });
-        res.status(200).json({ message: 'successfully deleted Product' });
+        const requesterEmail = req.decodedEmail;
+        if (requesterEmail) {
+            const requesterAccount = await User.findOne({
+                email: requesterEmail,
+            });
+            if (requesterAccount.role === 'admin' || requesterAccount.role === 'administer') {
+                await Product.deleteOne({ id: req.params.id });
+                res.status(200).json({ message: 'successfully deleted Product' });
+            } else {
+                res.status(500).json({ message: 'Only Admin can delete product' });
+            }
+        }
     } catch (error) {
         res.status(500).json(error.message)
     }
-
-
 }
